Seperate v2 cli logic into mod and ohttp components
This allows for easier review of the logic in the payjoin-cli as it
relates to the v2 relay fallback logic.

This change come in before the deprecation and removal of
_danger-local-https for even simpler logic here.
